--- a/CHANGELOG.rst
+++ b/CHANGELOG.rst
@@ -28,3 +28,5 @@ master
 * Added test for Django 3 using Python 3.7 in tox.ini
 * Improved performance of fake_text
 * Improved performance of BaseAnonymizer.patch_object
+* Improved performance of fake_username
+* Removed range_range argument from fake_username (backwards incompatible)