--- a/CHANGELOG.rst
+++ b/CHANGELOG.rst
@@ -29,4 +29,4 @@ master
 * Improved performance of fake_text
 * Improved performance of BaseAnonymizer.patch_object
 * Improved performance of fake_username
-* Removed range_range argument from fake_username (backwards incompatible)
+* Removed rand_range argument from fake_username (backwards incompatible)